Fixing after first review
