+++ next-app/src/app/administration/analytics/page.tsx
@@ -5,6 +5,7 @@ import { Input } from "@/components/ui/input";
 import NewUsersGraph from "@/components/administration/analytics/NewUsersGraph";
 import TotalLiquidityGraph from "@/components/administration/analytics/TotalLiquidityGraph";
 import TotalTradingVolumeGraph from "@/components/administration/analytics/TotalTradingVolumeGraph";
+import { SearchIcon } from "@/components/Icons";
 
 const AnalyticsPage: React.FC = () => {
   return (
@@ -34,25 +35,5 @@ const AnalyticsPage: React.FC = () => {
   </>
   );
 };
-
-function SearchIcon(props: React.JSX.IntrinsicAttributes & React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="11" cy="11" r="8" />
-      <path d="m21 21-4.3-4.3" />
-    </svg>
-  );
-}
 
 export default AnalyticsPage;
