+++ next-app/src/app/administration/platform/page.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { Input } from "@/components/ui/input";
 import PlatformFee from "@/components/administration/platform/PlatformFee";
 import BanUserAddress from "@/components/administration/platform/BanUserAddress";
+import { SearchIcon } from "@/components/Icons";
 
 const PlatformPage: React.FC = () => {
   return (
@@ -33,25 +34,5 @@ const PlatformPage: React.FC = () => {
     </>
   );
 };
-
-function SearchIcon(props: React.JSX.IntrinsicAttributes & React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="11" cy="11" r="8" />
-      <path d="m21 21-4.3-4.3" />
-    </svg>
-  );
-}
 
 export default PlatformPage;
