+++ next-app/src/app/administration/tokens/page.tsx
@@ -2,6 +2,7 @@
 import React from "react";
 import { Input } from "@/components/ui/input";
 import Airdrop from "@/components/administration/tokens/Airdrop";
+import { SearchIcon } from "@/components/Icons";
 
 const TokensPage: React.FC = () => {
   return (
@@ -29,25 +30,5 @@ const TokensPage: React.FC = () => {
     </>
   );
 };
-
-function SearchIcon(props: React.JSX.IntrinsicAttributes & React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="11" cy="11" r="8" />
-      <path d="m21 21-4.3-4.3" />
-    </svg>
-  );
-}
 
 export default TokensPage;
