+++ next-app/src/components/administration/analytics/NewUsersGraph.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { LineChart } from "@/components/Charts";
 import {
   Card,
   CardHeader,
@@ -7,98 +8,21 @@ import {
   CardDescription,
   CardContent,
 } from "@/components/ui/card";
-import { ResponsiveLine } from "@nivo/line";
-import { JSX, ClassAttributes, HTMLAttributes } from "react";
 
 const NewUsersGraph = () => {
   return (
-    <div>
-      <Card>
-        <CardHeader>
-          <CardTitle>New Users</CardTitle>
-          <CardDescription>
-            The number of new users joining the platform.
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <LineChart className="aspect-[9/4]" />
-        </CardContent>
-      </Card>
-    </div>
+    <Card>
+      <CardHeader>
+        <CardTitle>New Users</CardTitle>
+        <CardDescription>
+          The number of new users joining the platform.
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        <LineChart className="aspect-[9/4]" />
+      </CardContent>
+    </Card>
   );
 };
-
-function LineChart(props: JSX.IntrinsicAttributes & ClassAttributes<HTMLDivElement> & HTMLAttributes<HTMLDivElement>) {
-  return (
-    <div {...props}>
-      <ResponsiveLine
-        data={[
-          {
-            id: "Desktop",
-            data: [
-              { x: "Jan", y: 43 },
-              { x: "Feb", y: 137 },
-              { x: "Mar", y: 61 },
-              { x: "Apr", y: 145 },
-              { x: "May", y: 26 },
-              { x: "Jun", y: 154 },
-            ],
-          },
-          {
-            id: "Mobile",
-            data: [
-              { x: "Jan", y: 60 },
-              { x: "Feb", y: 48 },
-              { x: "Mar", y: 177 },
-              { x: "Apr", y: 78 },
-              { x: "May", y: 96 },
-              { x: "Jun", y: 204 },
-            ],
-          },
-        ]}
-        margin={{ top: 10, right: 10, bottom: 40, left: 40 }}
-        xScale={{
-          type: "point",
-        }}
-        yScale={{
-          type: "linear",
-        }}
-        axisTop={null}
-        axisRight={null}
-        axisBottom={{
-          tickSize: 0,
-          tickPadding: 16,
-        }}
-        axisLeft={{
-          tickSize: 0,
-          tickValues: 5,
-          tickPadding: 16,
-        }}
-        colors={["#2563eb", "#e11d48"]}
-        pointSize={6}
-        useMesh={true}
-        gridYValues={6}
-        theme={{
-          tooltip: {
-            chip: {
-              borderRadius: "9999px",
-            },
-            container: {
-              fontSize: "12px",
-              textTransform: "capitalize",
-              borderRadius: "6px",
-            },
-          },
-          grid: {
-            line: {
-              stroke: "#f3f4f6",
-            },
-          },
-        }}
-        role="application"
-      />
-    </div>
-  );
-}
 
 export default NewUsersGraph;
