+++ next-app/src/components/administration/platform/PlatformFee.tsx
@@ -1,5 +1,6 @@
 "use client";
-import { JSX, SVGProps, useState } from "react";
+
+import { useState } from "react";
 import {
   Card,
   CardHeader,
@@ -10,11 +11,12 @@ import {
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
+import { CircleCheckIcon } from "@/components/Icons";
 
 const PlatformFee = () => {
   const [feePercentage, setFeePercentage] = useState(3);
   const [showConfirmation, setShowConfirmation] = useState(false);
-  const handleFeePercentageChange = (e: { target: { value: string; }; }) => {
+  const handleFeePercentageChange = (e: { target: { value: string } }) => {
     const value = parseFloat(e.target.value);
     if (isNaN(value) || value < 0 || value > 10) {
       return;
@@ -69,25 +71,5 @@ const PlatformFee = () => {
     </div>
   );
 };
-
-function CircleCheckIcon(props: JSX.IntrinsicAttributes & SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <circle cx="12" cy="12" r="10" />
-      <path d="m9 12 2 2 4-4" />
-    </svg>
-  );
-}
 
 export default PlatformFee;
